Split the large xfr module into separate sub-modules.
